pyproject.toml
[project]
name = "password-store"
version = "0.1.0"
description = "password-store"
dependencies = [
  "ruamel.yaml",
]

[project.scripts]
password-store = "password_store:main"
